+++ plugins/catalog-backend-module-gcp/src/providers/GkeEntityProvider.ts
@@ -29,9 +29,15 @@ import {
   ANNOTATION_KUBERNETES_API_SERVER,
   ANNOTATION_KUBERNETES_API_SERVER_CA,
   ANNOTATION_KUBERNETES_AUTH_PROVIDER,
+  ANNOTATION_KUBERNETES_DASHBOARD_APP,
+  ANNOTATION_KUBERNETES_DASHBOARD_PARAMETERS,
 } from '@backstage/plugin-kubernetes-common';
 import { Config } from '@backstage/config';
 import { SchedulerService } from '@backstage/backend-plugin-api';
+import {
+  ANNOTATION_LOCATION,
+  ANNOTATION_ORIGIN_LOCATION,
+} from '@backstage/catalog-model';
 
 /**
  * Catalog provider to ingest GKE clusters
@@ -120,10 +126,16 @@ export class GkeEntityProvider implements EntityProvider {
 
   private clusterToResource(
     cluster: container.protos.google.container.v1.ICluster,
+    project: string,
   ): DeferredEntity | undefined {
     const location = `${this.getProviderName()}:${cluster.location}`;
 
-    if (!cluster.name || !cluster.selfLink || !location || !cluster.endpoint) {
+    if (
+      !cluster.name ||
+      !cluster.selfLink ||
+      !cluster.endpoint ||
+      !cluster.location
+    ) {
       this.logger.warn(
         `ignoring partial cluster, one of name=${cluster.name}, endpoint=${cluster.endpoint}, selfLink=${cluster.selfLink} or location=${cluster.location} is missing`,
       );
@@ -142,8 +154,14 @@ export class GkeEntityProvider implements EntityProvider {
             [ANNOTATION_KUBERNETES_API_SERVER_CA]:
               cluster.masterAuth?.clusterCaCertificate || '',
             [ANNOTATION_KUBERNETES_AUTH_PROVIDER]: 'google',
-            'backstage.io/managed-by-location': location,
-            'backstage.io/managed-by-origin-location': location,
+            [ANNOTATION_KUBERNETES_DASHBOARD_APP]: 'gke',
+            [ANNOTATION_LOCATION]: location,
+            [ANNOTATION_ORIGIN_LOCATION]: location,
+            [ANNOTATION_KUBERNETES_DASHBOARD_PARAMETERS]: JSON.stringify({
+              projectId: project,
+              region: cluster.location,
+              clusterName: cluster.name,
+            }),
           },
           name: cluster.name,
           namespace: 'default',
@@ -172,18 +190,22 @@ export class GkeEntityProvider implements EntityProvider {
     };
   }
 
-  private async getClusters(): Promise<
-    container.protos.google.container.v1.ICluster[]
-  > {
+  private async getClusters(): Promise<DeferredEntity[]> {
     const clusters = await Promise.all(
       this.gkeParents.map(async parent => {
+        const project = parent.split('/')[1];
         const request = {
           parent: parent,
         };
         const [response] = await this.clusterManagerClient.listClusters(
           request,
         );
-        return response.clusters?.filter(this.filterOutUndefinedCluster) ?? [];
+        return (
+          response.clusters
+            ?.filter(this.filterOutUndefinedCluster)
+            .map(c => this.clusterToResource(c, project))
+            .filter(this.filterOutUndefinedDeferredEntity) ?? []
+        );
       }),
     );
     return clusters.flat();
@@ -196,18 +218,14 @@ export class GkeEntityProvider implements EntityProvider {
 
     this.logger.info('Discovering GKE clusters');
 
-    let clusters: container.protos.google.container.v1.ICluster[];
+    let resources: DeferredEntity[];
 
     try {
-      clusters = await this.getClusters();
+      resources = await this.getClusters();
     } catch (e) {
       this.logger.error('error fetching GKE clusters', e);
       return;
     }
-    const resources =
-      clusters
-        .map(c => this.clusterToResource(c))
-        .filter(this.filterOutUndefinedDeferredEntity) ?? [];
 
     this.logger.info(
       `Ingesting GKE clusters [${resources