+++ plugins/catalog-react/src/components/MissingAnnotationEmptyState/index.ts
@@ -13,11 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-export type {
-  AppNode,
-  AppNodeEdges,
-  AppNodeInstance,
-  AppNodeSpec,
-} from './types';
-export { createAppGraph } from './createAppGraph';
+export * from './MissingAnnotationEmptyState';