+++ plugins/github-actions/src/components/Router.tsx
@@ -16,13 +16,15 @@
 
 import React from 'react';
 import { Entity } from '@backstage/catalog-model';
-import { useEntity } from '@backstage/plugin-catalog-react';
+import {
+  useEntity,
+  MissingAnnotationEmptyState,
+} from '@backstage/plugin-catalog-react';
 import { Routes, Route } from 'react-router-dom';
 import { buildRouteRef } from '../routes';
 import { WorkflowRunDetails } from './WorkflowRunDetails';
 import { WorkflowRunsTable } from './WorkflowRunsTable';
 import { GITHUB_ACTIONS_ANNOTATION } from './getProjectNameFromEntity';
-import { MissingAnnotationEmptyState } from '@backstage/core-components';
 
 /** @public */
 export const isGithubActionsAvailable = (entity: Entity) =>