+++ plugins/bitbucket-cloud-common/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-bitbucket-cloud-common
 
+## 0.2.14-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/integration@1.7.2-next.0
+
 ## 0.2.13
 
 ### Patch Changes