+++ plugins/cicd-statistics/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-cicd-statistics
 
+## 0.1.28-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.9.0-next.1
+  - @backstage/catalog-model@1.4.3
+  - @backstage/core-plugin-api@1.8.0-next.0
+
 ## 0.1.28-next.0
 
 ### Patch Changes