+++ plugins/gcalendar/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-gcalendar
 
+## 0.3.20-next.1
+
+### Patch Changes
+
+- 62b5922916: Internal theme type updates
+- Updated dependencies
+  - @backstage/core-components@0.13.8-next.1
+  - @backstage/core-plugin-api@1.8.0-next.0
+  - @backstage/errors@1.2.3
+  - @backstage/theme@0.4.4-next.0
+
 ## 0.3.20-next.0
 
 ### Patch Changes