+++ plugins/home-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-home-react
 
+## 0.1.5-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.13.8-next.1
+  - @backstage/core-plugin-api@1.8.0-next.0
+
 ## 0.1.5-next.0
 
 ### Patch Changes