+++ plugins/techdocs-addons-test-utils/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/plugin-techdocs-addons-test-utils
 
+## 1.0.23-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog@1.15.0-next.1
+  - @backstage/plugin-search-react@1.7.2-next.1
+  - @backstage/integration-react@1.1.21-next.1
+  - @backstage/plugin-techdocs@1.8.1-next.1
+  - @backstage/core-components@0.13.8-next.1
+  - @backstage/core-app-api@1.11.1-next.0
+  - @backstage/core-plugin-api@1.8.0-next.0
+  - @backstage/test-utils@1.4.5-next.0
+  - @backstage/theme@0.4.4-next.0
+  - @backstage/plugin-techdocs-react@1.1.13-next.1
+
 ## 1.0.23-next.0
 
 ### Patch Changes